Update nobackuptag.py
This script checks if a tag called backup exists for ec2 instances, ebs volumes, dynamodb tables & rds instances. It uses AWS secrets manager to do this across multiple accounts in a for loop & another for loop so that it is cross region. It collates this information into a CSV & uploads the information into an s3 bucket.